requirements.txt
click==8.0.1
termcolor
xmltodict
progress
requests
pyyaml